pyproject.toml
[tool.ruff]
line-length = 120
lint.select = ["E", "F", "W", "I"]
lint.ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
[project]
name = "supertd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pytest",
    "pyyaml>=6.0.2",
    "uv",
]
